Feature: FocusScreen 가로 이미지 적용
- graphicsLayer의 rotationZ 90도 적용
- 세로, 가로 이미지에 따른 계산법 차등 적용
