+++ feature/detail/src/main/java/com/kolown/porring/feature/detail/FocusScreen.kt
@@ -1,0 +1,130 @@
+package com.kolown.porring.feature.detail
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.gestures.detectTransformGestures
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableFloatStateOf
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clipToBounds
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.onSizeChanged
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.IntSize
+import com.kolown.porring.core.designsystem.component.PorringIconButton
+import com.kolown.porring.core.designsystem.component.PorringTopAppBar
+import com.kolown.porring.core.designsystem.ui.theme.BackgroundDark
+import com.kolown.porring.core.ui.component.CoilImage
+
+@Composable
+internal fun FocusScreen(
+    postUrl: String = "",
+    imageRatio: Float = 4f / 5f,
+    onDismiss: () -> Unit = {},
+) {
+    val isHorizontalImage = imageRatio > 1f
+    val imageRotation = if (isHorizontalImage) 90f else 0f
+    val appBarAlignment = if (isHorizontalImage) Alignment.BottomCenter else Alignment.TopCenter
+
+    val minScale = if (isHorizontalImage) 1.25f else 1f
+    val maxScale = 2f
+    var scale by remember { mutableFloatStateOf(minScale) }
+    var offset by remember { mutableStateOf(Offset.Zero) }
+    var boxSize by remember { mutableStateOf(IntSize.Zero) }
+    var imageSize by remember { mutableStateOf(IntSize.Zero) }
+
+
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(BackgroundDark)
+            .onSizeChanged { boxSize = it }
+    ) {
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .align(Alignment.Center)
+                .clipToBounds()
+                .pointerInput(Unit) {
+                    detectTransformGestures { _, pan, zoom, _ ->
+                        scale = (scale * zoom).coerceIn(minScale, maxScale)
+
+                        val maxOffset = calcMaxOffset(
+                            isHorizontalImage = isHorizontalImage,
+                            scale = scale,
+                            boxSize = boxSize,
+                            imageSize = imageSize
+                        )
+
+                        offset = Offset(
+                            x = (offset.x + pan.x).coerceIn(-maxOffset.x, maxOffset.x),
+                            y = (offset.y + pan.y).coerceIn(-maxOffset.y, maxOffset.y)
+                        )
+                    }
+                }
+        ) {
+            CoilImage(
+                modifier = Modifier
+                    .align(Alignment.Center)
+                    .onSizeChanged { imageSize = it }
+                    .graphicsLayer(
+                        scaleX = scale,
+                        scaleY = scale,
+                        translationX = offset.x,
+                        translationY = offset.y,
+                        rotationZ = imageRotation
+                    ),
+                imageUrl = postUrl,
+                imageRatio = imageRatio,
+                isRipple = false
+            )
+        }
+
+        PorringTopAppBar(
+            modifier = Modifier.align(appBarAlignment),
+            trailingIcon = {
+                PorringIconButton(
+                    icon = Icons.Default.Close,
+                    onClick = onDismiss,
+                    contentDescription = stringResource(R.string.string_end_mode),
+                    color = Color.White
+                )
+            }
+        )
+    }
+}
+
+fun calcMaxOffset(
+    isHorizontalImage: Boolean,
+    scale: Float,
+    boxSize: IntSize,
+    imageSize: IntSize
+): Offset {
+    val scaledImageWidth =
+        if (isHorizontalImage) imageSize.height * scale else imageSize.width * scale
+    val scaledImageHeight =
+        if (isHorizontalImage) imageSize.width * scale else imageSize.height * scale
+
+    val maxOffsetX = ((scaledImageWidth - boxSize.width) / 2f).coerceAtLeast(0f)
+    val maxOffsetY = ((scaledImageHeight - boxSize.height) / 2f).coerceAtLeast(0f)
+
+    return Offset(maxOffsetX, maxOffsetY)
+}
+
+@Preview(showBackground = true, backgroundColor = 0x000000)
+@Composable
+private fun FocusScreenPreview() {
+    FocusScreen()
+}
